requirements.txt
boto==2.45.0
bz2file==0.98
cloudpickle==0.2.2
cycler==0.10.0
cymem==1.31.2
Cython==0.25.2
cytoolz==0.8.2
dask==0.12.0
decorator==4.0.10
flexmock==0.10.2
funcsigs==1.0.2
gensim==0.13.4.1
h5py==2.6.0
Keras==1.2.0
matplotlib==1.5.3
mock==2.0.0
murmurhash==0.26.4
networkx==1.11
nltk==3.2.1
numpy==1.11.3
pathlib==1.0.1
pbr==1.10.0
Pillow==3.4.2
plac==0.9.6
preshed==0.46.4
protobuf==3.1.0.post1
pycocotools==2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.13.0
scikit-image==0.12.3
scipy==0.18.1
semver==2.7.5
six==1.10.0
smart-open==1.4.0
spacy==1.6.0
sputnik==0.9.3
tensorflow-gpu==0.12.1
Theano==0.8.2
thinc==6.2.0
toolz==0.8.2
tqdm==4.11.2
ujson==1.35